day 10 and 11 BankManageValidation Done

+++ Core_Java/Day10/BankManage/src/com/bank/core/BankAccount.java
@@ -21,7 +21,7 @@ public class BankAccount {
 		this.accountNumber = accountNumber;
 	}
 
-	public double getAccountNumber() {
+	public int getAccountNumber() {
 		return accountNumber;
 	}
